backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
python-dotenv
python-multipart
passlib[bcrypt]
python-jose
pydantic
google-genai
faiss-cpu
pandas